backend/requirements.txt


fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pillow>=10.0.0
python-docx>=1.1.0
markdown>=3.5.1
pandas>=2.1.4
openpyxl>=3.1.2
supabase==2.7.0
resend>=0.7.0
python-dotenv>=1.0.0
lxml>=5.0.0
xmltodict>=0.13.0
pypdf>=4.0.0
pymupdf>=1.23.0
pdf2docx>=0.5.6
docx2pdf>=0.1.8
pytesseract>=0.3.10
